Refactor step 3/3: extract pipeline.py and make VMR+_1.7.14.py a thin entry point
- vmrplus/pipeline.py: _process_single_task, run_parallel_pipeline, main() (the
  old __main__ block wrapped as a function; family_markers/genus_name initialised
  at its top, previously module-level).
- VMR+_1.7.14.py: now a 40-line entry point (verbatim argparse block + main(args)).
  Invocation unchanged: python3 "VMR+_1.7.14.py" -config <ini> -thr <N>.
- Only non-verbatim edits vs original: shared singletons referenced as
  ncbi._ncbi_limiter / ncbi._entrez_cache (8 sites). 47/48 top-level defs proven
  byte-identical via inspect.getsource; -v/-h/no-args/--generate-config byte-equal
  to the original; 21 tests pass; ruff clean.
- tests import helpers from vmrplus.* now; conftest.py puts the package on sys.path.
- ruff.toml ignores F841 (pre-existing dead locals exposed by wrapping main()).
