+++ conftest.py
@@ -1,0 +1,1 @@
+# Ensures the project root is on sys.path so `import vmrplus` works under pytest.
